+++ src/app/components/export-dialog/export-dialog.component.spec.ts
@@ -1,4 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 
 import { ExportDialogComponent } from './export-dialog.component';
 
@@ -8,7 +9,13 @@ describe('ExportDialogComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [ExportDialogComponent]
+      imports: [ExportDialogComponent],
+      providers: [
+        {
+          provide: MAT_DIALOG_DATA,
+          useValue: { teiTranscriptions: false },
+        },
+      ],
     })
     .compileComponents();