+++ src/app/components/translate-description-dialog/translate-description-dialog.component.spec.ts
@@ -1,4 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 
 import { TranslateDescriptionDialogComponent } from './translate-description-dialog.component';
 
@@ -8,7 +9,16 @@ describe('TranslateDescriptionDialogComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [TranslateDescriptionDialogComponent]
+      imports: [TranslateDescriptionDialogComponent],
+      providers: [
+        {
+          provide: MAT_DIALOG_DATA,
+          useValue: {
+            descriptions: [{ language: 'en' }],
+            activeDescriptionIndex: 0,
+          },
+        },
+      ],
     })
     .compileComponents();